# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Loadbyton API
  version: 1.0.0
  description: UAE road freight marketplace — jobs, bids, escrow, payouts
servers:
  - url: http://localhost:4000
    description: Local
  - url: https://api.loadbyton.com
    description: Production (proxied from https://loadbyton.com/api/* per vercel.json)
...
paths:
...
  /api/stripe/connect/onboard:
    post:
      tags: [payments]
      summary: Carrier Connect onboarding — create Express account and account link
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: account link }
        '502': { description: provider error }
  /api/stripe/connect/status:
    get:
      tags: [payments]
      summary: Carrier Connect status — charges_enabled, payouts_enabled
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: ok }
  /api/jobs/{id}/pay:
    post:
      tags: [payments]
      summary: Create payment intent for escrow (shipper)
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: payment intent }
  /api/jobs/{id}/release-payout:
    post:
      tags: [payments]
      summary: Release payout via Stripe transfer (HSM multi-sig 2-of-3 when configured)
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: transfer }
        '403': { description: HSM required }
...
  /api/webhooks/stripe:
    post:
      tags: [payments]
      summary: Stripe webhook (raw body + stripe-signature verified, idempotent via payment_webhook_events)
      responses:
        '200': { description: ack }
        '400': { description: bad signature or unparseable payload }
  /api/webhooks/stripe/mock-confirm:
    post:
      tags: [payments]
      summary: Admin-only test webhook — simulates a successful Stripe payment (test/mock payment mode only)
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [processorPaymentRef]
              properties:
                processorPaymentRef: { type: string }
      responses:
        '200': { description: ok }
        '403': { description: not available outside test/mock payment mode }
components:
...
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: lb_session }